% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{no_assumption_bound}
\alias{no_assumption_bound}
\title{Calculate No Assumptions bounds
Main function of \code{otsbounds}. Calculates Optimal Treatment Selection bounds}
\usage{
no_assumption_bound(ps, y1, y0, beta, alpha)
}
\arguments{
\item{ps}{prognostic score}

\item{y1}{Observed effect on treatment 1}

\item{y0}{Observed effect on treatment 0}

\item{beta}{The lowest possible potential outcome}

\item{alpha}{The highest possible potential outcome}
}
\description{
Calculate No Assumptions bounds
Main function of \code{otsbounds}. Calculates Optimal Treatment Selection bounds
}
\examples{
ps = c(0.1, 0.2, 0.1, 0.4)
obs = data.frame(a1 = c(1, 2, 3, 4), a2 = c(4, 5, 6, 7))
a = 0
}
