% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{ots_bound}
\alias{ots_bound}
\title{Calculate OTS bounds
Main function of \code{otsbounds}. Calculates Optimal Treatment Selection bounds}
\usage{
ots_bound(ps, obs, a)
}
\arguments{
\item{ps}{prognostic score}

\item{obs}{Observed effect, e.g. observed mean}

\item{a}{The lowest bound for the potential outcome}
}
\description{
Calculate OTS bounds
Main function of \code{otsbounds}. Calculates Optimal Treatment Selection bounds
}
\examples{
ps = c(0.1, 0.2, 0.1, 0.4)
obs = data.frame(a1 = c(1, 2, 3, 4), a2 = c(4, 5, 6, 7))
a = 0
}
